Split generate initial job helpers
